+++ src/components/MyPage/CustomerCenter.js
@@ -16,7 +16,7 @@ export default function CustomerCenter() {
           }}
         >
           <Col xs={{ span: 3 }}>
-            <Link to="/mypage">
+            <Link to="/mypage/buy">
               <img
                 style={{
                   width: "32px",
